voice-agent/requirements.txt
PyAudio==0.2.14
websockets==12.0
janus==1.0.0
python-dotenv
fastapi
uvicorn
google-cloud-firestore
firebase_admin